Fix - Database connectivity and remove .env file

+++ src/main/java/edu/icet/lms/dto/StudentDto.java
@@ -13,7 +13,7 @@ public class StudentDto {
     private String id;
     private String firstName;
     private String lastName;
-    private Integer nic;
+    private Double nic;
     private String address;
     private String phone;
     private String city;
